% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_G.R
\name{calc_G}
\alias{calc_G}
\title{Calculate Geometric means of columns}
\usage{
calc_G(
  x = "raDeltaAbund",
  y = c("TAXON_ID", "COMMON_NAME", "SCIENTIFIC_NAME", "DIVNAME", "EPBC_ACT_STATUS",
    "FFG_ACT_STATUS", "CombThreshold", "Baseline", "NoLessthanThreshhold",
    "LastLessThanThreshold")
)
}
\arguments{
\item{x}{data.frame containing columns with numeric values}

\item{y}{vector of names of columns to exclude from calculation}
}
\value{
vector numeric or NA if not all values in column are > 0
}
\description{
Calculate Geometric means of columns
}
\details{
Calculates geometric means of numeric columns of dataframe
}
