% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growth_stages.R
\name{calc_growth_stages}
\alias{calc_growth_stages}
\title{Summarise area by growth stage.}
\usage{
calc_growth_stages(
  myFHAnalysis = FHAnalysis,
  myCropRasters = cropRasters,
  myAllCombs = allCombs,
  OutputRasters = makeGSRasters,
  myResultsDir = ResultsDir
)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes
created by function fhProcess()}

\item{myCropRasters}{object containing cell indices and values of input
rasters cropped to area of interest created by cropToOutput()}

\item{myAllCombs}{list made by function calc_U_AllCombs}

\item{OutputRasters}{logical whether to output rasters of TFI status for each
year}

\item{myResultsDir}{path of directory where results will be written usually
generated  by FAME script}
}
\value{
list of two data.frames giving area of each growth stage grouped by
EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME, FIRE_REGION_NAME, DELWP_REGION.
Optional byproduct raster giving growth
stages for each year in RASTER attribute table.
\itemize{
\item GS_Summary_wide Wide format table summarises area by Growth Stage and SEASON
\item GS_Summary_long Long format table summarises area by Growth Stage and SEASON
}
}
\description{
Summarise area by growth stage.
}
\details{
Generates wide and long format summary of area for each EFG and
season grouped by EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME, FIRE_REGION_NAME,
DELWP_REGION.
}
