% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFI_functions.R
\name{calc_TFI_2}
\alias{calc_TFI_2}
\title{Main Tolerable fire interval (TFI) status calculation}
\usage{
calc_TFI_2(
  FHanalysis,
  U_AllCombs_TFI = myAllCombs$U_AllCombs_TFI,
  Index_AllCombs = myAllCombs$Index_AllCombs,
  TFI_LUT,
  OutputRasters = makeTFIRasters
)
}
\arguments{
\item{FHanalysis}{list containing all the fire history spatial attributes
created by function fhProcess()}

\item{U_AllCombs_TFI}{data.table giving all combinations of cell values from
the input rasters for the FAME analysis made in function calcU_All_Combs()}

\item{Index_AllCombs}{integer index mapping U_AllCombs_TFI to raster cells
made in function calcU_All_Combs()}

\item{TFI_LUT}{data.frame Lookup table from EFG for
"MIN_LO_TFI","MIN_HI_TFI","MAX_TFI","EFG_NAME", read form settings}

\item{OutputRasters}{logical whether to uptut rasters of TFI status for each
year}
}
\value{

}
\description{
Main Tolerable fire interval (TFI) status calculation
}
\details{
Calculates where each cell is currently at below MinTFI or above
MAX_TFI returns the per cell and long table summarised by multiple admin
units and evc
}
