% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBBTFI_2.R
\name{calcBBTFI_2}
\alias{calcBBTFI_2}
\title{Calculate area BBTFI and BBTFI rasters}
\usage{
calcBBTFI_2(
  myFHAnalysis = FHAnalysis,
  myAllCombs = allCombs,
  makeBBTFIrasters = makeBBTFIrasters,
  myResultsDir = NULL
)
}
\arguments{
\item{myFHAnalysis}{list of Fhire history analysis components
created by function fhProcess()}

\item{myAllCombs}{list made by function calc_U_AllCombs}

\item{makeBBTFIrasters}{logical whether or not to export
rasters for BBTFI to disk}

\item{myResultsDir}{path of directory where results will be written usually
generated  by FAME script}
}
\value{
list containing:
\itemize{
\item the date sequence matrix for each cell of the raster
\item the EFG TFI Lookup for each cell of the raster
\item the raster resolution used.
\item Optionally outputs rasters of BBTFI to disk if makeBBTFIrasters==TRUE.
}
}
\description{
Calculate area BBTFI and BBTFI rasters
}
\details{
Calculate summary  area burned below TFI BBTFI for each SEASON in analysis
(accommodating  Hi and Lo fire intensity of first burn to determine TFI)
and cumulative area BBTFI.  Also optionally outputs rasters mapping areas BBTFI
}
