% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_abundance_functions.R
\name{calcDeltaAbund}
\alias{calcDeltaAbund}
\title{Summary of changes in relative abundance}
\usage{
calcDeltaAbund(
  SpYearSumm = SpYearSummWide,
  FHanalysis,
  myBaseline,
  ResultsDir,
  HDMSpp_NO,
  TaxonList
)
}
\arguments{
\item{SpYearSumm}{data.frame output by function makeSppYearSum2()}

\item{FHanalysis}{list containing all the fire history spatial attributes created by function fhProcess}

\item{myBaseline}{integer single SEASON or sequence of SEASons used to create the baseline relative species abundance for comparison of change}

\item{ResultsDir}{path of directory where results will be written usually generated by FAME script}

\item{HDMSpp_NO}{path of directory where results will be written usually generated by FAME script}

\item{TaxonList}{data.frame of species attributes ( read from default or user provided .csv)}
}
\value{
data frame wide format summary chance in relative abundance of species SEASON
}
\description{
Summary of changes in relative abundance
}
\details{
Calculates the change in relative abundance compared to a baseline SEASON or mean of SEASONS
}
