% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_U_AllCombs.r
\name{calcU_All_Combs}
\alias{calcU_All_Combs}
\title{Calculate all combinations of input raster values}
\usage{
calcU_All_Combs(FHAnalysis = FHAnalysis, cropRasters = cropRasters)
}
\arguments{
\item{FHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess()}

\item{cropRasters}{list of rasters and indices and cell values created by function cropNAborder()}
}
\value{
list of:
\
}
\description{
Calculate all combinations of input raster values
}
