% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropNAborder.R, R/cropNAborder_dev.R
\name{cropNAborder}
\alias{cropNAborder}
\title{also creates some rasters cropped to correct extent for instance for region and EFG
also gets indices of cells in raster of same extent as crop to the shape provided}
\usage{
cropNAborder(
  REG_NO = 7,
  myRasterRes = RasterRes,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters"
)

cropNAborder(
  REG_NO = 7,
  myRasterRes = RasterRes,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters"
)
}
\arguments{
\item{REG_NO}{integer DELWP fire region number 1:6 ,99 for Statewide analysis,  or 7 for ad hoc boundary polygon default =7 (see look up table REG_LUT for values)}

\item{myRasterRes}{numeric raster resolution of the analysis in metres ( usually set in settings file or shiny app)}

\item{PUBLIC_LAND_ONLY}{Logical TRUE/FALSE}

\item{myPoly}{default clipPoly sf polygon data frame of LF_REGIONs (default) or ad hoc polygon - used in conjunction with REG_NO}

\item{generalRasterDir}{relative path to directory containing rasters of DELWP FIRE_REG, DELWP REGION, EFG, PUBLIC LAND (PLM_GEN)}
}
\value{
A list containing:
\itemize{
\item Raster  raster cropped to extent of area of interest,
\item IDX integer vector cell numbers of cells in the cropped raster
\item clipIDX integer vector cell numbers only for cells with the input polygon
\item EFG integer vector EFG values for cells within clipped area
\item RGN integer vector Fire Region numbers for cells within clipped area
\item DELWP integer vector DELWP Region numbers for cells within clipped area
\item PLM logical for cells within clipped area
}

A list containing:
\itemize{
\item Raster  raster cropped to extent of area of interest,
\item IDX integer vector cell numbers of cells in the cropped raster
\item clipIDX integer vector cell numbers only for cells with the input polygon
\item EFG integer vector EFG values for cells within clipped area
\item RGN integer vector Fire Region numbers for cells within clipped area
\item DELWP integer vector DELWP Region numbers for cells within clipped area
\item PLM logical for cells within clipped area
}
}
\description{
also creates some rasters cropped to correct extent for instance for region and EFG
also gets indices of cells in raster of same extent as crop to the shape provided

also creates some rasters cropped to correct extent for instance for region and EFG
also gets indices of cells in raster of same extent as crop to the shape provided
}
