% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropToOutput.R
\name{cropToOutput}
\alias{cropToOutput}
\title{Creates an output list containing vectors of input rasters values, definintion of an empty
output raster (by extent and resoloution) cropped to the clipPolygon provided
also gets indices of cells in raster of same extent as crop to the shape provided
from input and output}
\usage{
cropToOutput(
  REG_NO = 7,
  myRasterRes = RasterRes,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters"
)
}
\arguments{
\item{REG_NO}{integer DELWP fire region number 1:6 ,99 for Statewide analysis,  or 7 for ad hoc boundary polygon default =7 (see look up table REG_LUT for values)}

\item{myRasterRes}{numeric raster resolution of the analysis in metres ( usually set in settings file or shiny app)}

\item{PUBLIC_LAND_ONLY}{Logical TRUE/FALSE}

\item{myPoly}{path of polygon file  LF_REGIONs (default) or ad hoc polygon - used in conjunction with REG_NO to crop the and output rasters.}

\item{generalRasterDir}{relative path to directory containing rasters of DELWP FIRE_REG, DELWP REGION, EFG, PUBLIC LAND (PLM_GEN)}
}
\value{
A list containing:
\itemize{
\item inCells integer vector cell numbers of cells in the inputRaster(s)
that correspond to myPoly
\item outCells integer vector cell numbers of cells in the outputRaster(s)
that correspond to the inputRaster
\item rasterDef an expression defining an empty raster with the correct extent
for the outputRaster
\item EFG integer vector EFG values for cells within clipped area
\item RGN integer vector Fire Region numbers for cells within clipped area
\item DELWP integer vector DELWP Region numbers for cells within clipped area
\item PLM logical for cells within clipped area
}
}
\description{
Creates an output list containing vectors of input rasters values, definintion of an empty
output raster (by extent and resoloution) cropped to the clipPolygon provided
also gets indices of cells in raster of same extent as crop to the shape provided
from input and output
}
