% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_spatvector.R
\name{ensure_spatvector}
\alias{ensure_spatvector}
\title{Ensure input is a SpatVector}
\usage{
ensure_spatvector(x)
}
\arguments{
\item{x}{A \code{terra::SpatVector} object or a character file path.}
}
\value{
A \code{terra::SpatVector}.
}
\description{
Accepts either a \code{terra::SpatVector} object or a file path that can be
read into a \code{SpatVector}. Returns a valid \code{SpatVector}, or stops with a
warning message if unsuccessful.
}
\examples{
library(terra)

v <- vect(system.file("ex/lux.shp", package = "terra"))
ensure_spatvector(v)

ensure_spatvector(system.file("ex/lux.shp", package = "terra"))

}
