% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAMEFMR-package.R
\docType{package}
\name{FAMEFMR-package}
\alias{FAMEFMR}
\alias{FAMEFMR-package}
\title{FAMEFMR: Functions For DEECA FAME Fire Analysis}
\description{
"FameFMR version including qgisbased fhProcess"
}
\author{
\strong{Maintainer}: Nevil Amos \email{nevil.amos@gmail.com} (\href{https://orcid.org/0000-0003-1271-4767}{ORCID})

}
\keyword{internal}
