% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhProcess_qgis.R
\name{fhProcess_qgis}
\alias{fhProcess_qgis}
\title{complete Fire History Processing using QGIS_process}
\usage{
fhProcess_qgis(
  firstFH,
  firstFHLayer = NULL,
  secondFH = NULL,
  secondFHLayer = NULL,
  clipShape = NULL,
  start.SEASON = NULL,
  end.SEASON = NULL,
  OtherAndUnknown = 2,
  max_interval = 0,
  validFIRETYPE = c("BURN", "BUSHFIRE", "OTHER", "UNKNOWN"),
  baseFire = baseFire,
  precsision = 0,
  JFMPSeason0 = NA,
  addJFMPplus2 = FALSE
)
}
\arguments{
\item{firstFH}{path to the input fire history geodatabase (".gdb" or ".gpkg")
or shapefile (".shp") usually  provided in settings}

\item{firstFHLayer}{if rawFH is a geodatabase the name of the layer
containing the Fire History if this is not provided and the raw FH is a
geodatabase or geopackage then the first layer in the file is returned.This
argument is available only in the r script version - it is not exposed in
the shinyApp where it is set to NULL}

\item{secondFH}{Second fire history to be combined with FH1 to make a fire
scenario same formats as for firstFH or NULL if there is only a single FH
used.}

\item{secondFHLayer}{if secondFH is a geodatabase the name of the layer
containing the Fire History if this is not provided and the raw FH is a
geodatabase or geopackage then the first layer in the file is returned.This
argument is available only in the r script version - it is not exposed in
the shinyApp where it is set to NULL}

\item{clipShape}{vector polygon(s)as file path or name of sf vector object in
layer providing the boundary of the area of interest for the analysis. The
fire history inputs and output will be clipped to this polygon layer.}

\item{start.SEASON}{integer First SEASON for which output is wanted (four
digit year as integer), if NUll then second season in in history is used
(cannot use first season because it has no interval, this may still fail if
there is no overlap)}

\item{end.SEASON}{integer Last SEASON required, if NA or less than max value
in fire historythen largest value in fire history scenario used.}

\item{OtherAndUnknown}{integer Value to use for cases where fire type is:
"OTHER" or "UNKNOWN" = NA, "BURN" = 1, "BUSHFIRE" = 2. NA = Fire excluded
from analysis. usually set in settings file}

\item{max_interval}{= NULL #integer number of years  maximum inter fire
interval after a HIGH fire for which subsequent fire is reassigned to HIGH
default is NULL in which case there is no reassignment}

\item{validFIRETYPE}{vector of valid names in the input FIRETYPE column in
the input fire history dataset(s), if the column contains NA or values not
on this list an error will occur}

\item{baseFire}{Default NULL otherwise four digit integer SEASON for fire
applied across the whole bounding box}

\item{JFMPSeason0}{The fire season  year 0 of the JFMP when a specific JFMP
analysis is being carried out, otherwise NA, default NA}

\item{addJFMPplus2}{whether to add a fire everywhere in JFMPSeason0 + 2 as
required for JFMP analysis, default FALSE}
}
\value{
A list containing: \itemize{
\item OutDF sf polygons dataframe containing all the fire history attributes
\item TimeSpan integer vector sequence of SEASONS to in the analysis output
\item YSFNames names of TSF years in output, needed by downstream functions
\item LBYNames names of LBY years in output, needed by downstream functions
\item LFTNames names of LBY years in output, needed by downstream functions }
}
\description{
complete Fire History Processing using QGIS_process
}
\details{
The function takes a shapefile or  geopackage layer of Fire history
containing polygons with two fields: FIRETYPE and SEASON Where polygons of
different FIRETYPE or SEASON overlap the function constructs unique
non-overlapping polygon of their intersections ( and non intersecting areas
) and attributes each polygon with sequential fire SEASON (SEAS01, SEAS02
...) and corresponding FIRETYPE (TYPE01,TYPE02 ...)

It then calculates all the intervals between sequential fires, and
Time Since fire (TSF) and Last Fire Type (LFT) and Last burnt year (LBY)
for each SEASON as defined in the input arguments, these values are append to
the output sf polygon dataframe.
}
