% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhProcessGDB.R
\name{fhProcessGDB}
\alias{fhProcessGDB}
\title{Main Fire History Fire Sequence analysis function GDB version needs testing more}
\usage{
fhProcessGDB(
  readGDB = FALSE,
  gdbPath = NULL,
  rawFH = "path of the rawFH file to use - a shapefile",
  start.SEASON = NULL,
  end.SEASON = NULL,
  OtherAndUnknown,
  validFIRETYPE
)
}
\arguments{
\item{readGDB}{Boolean TRUE if the input is a featureclass in a geodatabase, not a shapefile. Default = False}

\item{gdbPath}{path to geodatabase containing rawFH if readGDB =  TRUE, otherwise NULL.}

\item{rawFH}{path to the input fire history shapefile usually  provided in settings. Alternatively the name of a feature class in gdb if readGDB =  TRUE}

\item{start.SEASON}{integer First SEASON for which output is wanted (four digit year as integer), if NUll then second season in in history is used (cannot use first season because it has no interval, this may still fail if there is no overlap)}

\item{end.SEASON}{integer Last SEASON required, if NULL then largest value in fire history scenario used}

\item{OtherAndUnknown}{integer Value to use for cases where fire type is: "OTHER" or "UNKNOWN" = NA, "BURN" = 1, "BUSHFIRE" = 2. NA = Fire excluded from analysis. usually set in settings file}

\item{validFIRETYPE}{character vector of valid FIRETYPE values for checking the input file , provided in settings file.}
}
\value{
A list containing:
\itemize{
\item OutDF sf polygons dataframe containing all the fire history attributes
\item TimeSpan integer vector sequence of SEASONS to in the analysis output
\item YSFNames names of  TSF years in output, needed by downstream functions
\item LBYNames names of  LBY years in output, needed by downstream functions
\item LFTNames names of  LBY years in output, needed by downstream functions
}
}
\description{
Main Fire History Fire Sequence analysis function GDB version needs testing more
}
\details{
The function takes a shapefile of Fire history contain polygons with two
fields FIRETYPE and SEASON Where polygons of different FIRETYPE or SEASON
overlap the function constructs unique non-overlapping polygon of their
intersections ( and non intersecting areas ) and attributes each polygon with
sequential fire SEASON (SEAS01, SEAS02 ...) and corresponding FIRETYPE
(TYPE01,TYPE02 ...)

It then calculates all the intervals between sequential fires, and Time Since
fire (TSF) and Last Fire Type (LFT) and Last burnt year (LBY) for each SEASON as defined in the input
arguments, these values are append to the output sf polygon dataframe.
}
