% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_fire_history.R
\name{generate_random_fire_history}
\alias{generate_random_fire_history}
\title{Generate Random Fire History Polygons}
\usage{
generate_random_fire_history(
  nPolygons = 10,
  bbox = sf::st_bbox(c(xmin = 2600000, xmax = 2620000, ymax = 2420000, ymin = 2400000)),
  maxPoints = 51,
  seasonRange = 1980:2025,
  crs = sf::st_crs(3111),
  validFIRETYPE = c("BURN", "BUSHFIRE", "UNKNOWN", "OTHER")
)
}
\arguments{
\item{nPolygons}{number of individual polygons to be generated}

\item{bbox}{bounding box within which the polygons are to be generated
including its crs is st_bbox format. default is a 20km square in Central
Eastern Victoria and epsg:3111 /VicGrid94 crs}

\item{maxPoints}{the maximum number of points used to define each polygon ( min value 11).}

\item{seasonRange}{vector of integer season values selected at random for polygon attribute SEASON}

\item{crs}{coordinate reference system  for the output set using sf_crs(),
default epsg:3111}

\item{validFIRETYPE}{vector of valid names in the input FIRETYPE column in
the input fire history dataset(s), if the column contains NA or values not
on this list an error will occur}
}
\value{
sf polygon
}
\description{
Function generates random fire history polygon dataset
}
\examples{
require(sf)
randomFH<-generate_random_fire_history(nPolygons=20,maxPoints=11)
plot(randomFH["SEASON"])
plot(randomFH["FIRETYPE"])

}
