% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{geoMean}
\alias{geoMean}
\title{Calculate Geometric mean}
\usage{
geoMean(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric or NA if values are not all > 0
}
\description{
Calculate Geometric mean
}
\details{
Calculates geometric mean of a vector of positive numeric values
copied from R package EnvStats https://cran.r-project.org/web/packages/EnvStats/index.html
}
