% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_fire_season}
\alias{get_fire_season}
\title{Get Fire Season of Given Date}
\usage{
get_fire_season(date)
}
\arguments{
\item{date}{A vector of dates, coercible to class \code{Date}.}
}
\value{
An integer vector giving the ending year of the fire season.
}
\description{
Returns the integer year of the fire season in which a date falls.
The fire season is assumed to run from \strong{1 July to 30 June}.
}
\details{
For example:
\itemize{
\item "2024-05-15" → 2024 (Fire Season 2023–24)
\item "2024-09-01" → 2025 (Fire Season 2024–25)
}
}
\examples{
get_fire_season("2024-05-15")   # returns 2024
get_fire_season("2024-09-01")   # returns 2025
get_fire_season(Sys.Date())     # today's fire season year

}
