% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfmpRASumm.R
\name{jfmpRASumm}
\alias{jfmpRASumm}
\title{Summary of relative abundance changes comparing JFMP results}
\usage{
jfmpRASumm(
  myDraftJfmpOut = rv$draftJfmpOut,
  myGrpSpYearSummLong = rv$grpSpYearSummLong,
  myTaxonList = rv$TaxonList,
  myStartBaseline = rv$startBaseline,
  myEndBaseline = rv$endBaseline
)
}
\arguments{
\item{myDraftJfmpOut}{data.frame of autoJFMP joined to alternative draft JFMPs mad by function joinDraftJFMP()}

\item{myGrpSpYearSummLong}{data.frame Long format relative abundances of species per PU and EFG and SEASON, created when calc_SpeciesRA() is run on FHAnalysis containing PU}

\item{myTaxonList}{data.frame of species attributes (read from default or
user provided .csv)}

\item{myStartBaseline}{first SEASON of years to use for calcuation RA baseline}

\item{myEndBaseline}{last SEASON of years to use for calcuation RA baseline}
}
\value{
list of two data.frames
\itemize{
\item  jfmpSppRaSumm Summary per species for each JFMP and NoJFMP of change relative to baseline, thresholds and summed relative abundance
\item  nBelowThreshHold count of number of species below threshold for each JFMP and NoJFMP in JFMPSeson0 + 4}
}
\description{
Summary of relative abundance changes comparing JFMP results
}
