% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_stage_functions.R
\name{makeGS_Summary}
\alias{makeGS_Summary}
\title{Summarise area by growth stage.}
\usage{
makeGS_Summary(
  FHanalysis,
  U_AllCombs_TFI = myAllCombs$U_AllCombs_TFI,
  Index_AllCombs = myAllCombs$Index_AllCombs
)
}
\arguments{
\item{FHanalysis}{list containing all the fire history spatial attributes
created by function fhProcess}

\item{U_AllCombs_TFI}{data.table giving all combinations of cell values from
the input rasters for the FAME anaysis}

\item{Index_AllCombs}{integer index mapping U_AllCombs_TFI to raster cells #'}
}
\value{
list of two data.frames grouped by EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME,
FIRE_REGION_NAME, DELWP_REGION
\itemize{
\item GS_Summary_wide Wide format table summarises area by Growth Stage and SEASON
\item GS_Summary_long Long format table summarises area by Growth Stage and SEASON
}
}
\description{
Summarise area by growth stage.
}
\details{
Generates wide and long format summary of area for each EFG and
season grouped by EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME, FIRE_REGION_NAME,
DELWP_REGION.
}
