% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHDMVals.R
\name{makeHDMVals}
\alias{makeHDMVals}
\title{Extract HDM values for relevant cells and resolution for use in RA calculations}
\usage{
makeHDMVals(
  myHDMSpp_NO = HDMSpp_NO,
  myCropRasters = cropRasters,
  RasterRes = myFHAnalysis$RasterRes
)
}
\arguments{
\item{myHDMSpp_NO}{vector of TAXON_IDs for species to be included in output}

\item{myCropRasters}{list of rasters and indices and cell values created by function cropNAborder()}

\item{RasterRes}{numeric raster resolution of the analysis in metres (225 or 75
usually set in settings file or shiny app)}
}
\description{
Extract HDM values for relevant cells and resolution for use in RA calculations
}
