% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeHDMValsfromRasters}
\alias{makeHDMValsfromRasters}
\title{Function makes a matrix of HDM values(1,NA) constrained to those cells that
are indexed in the cropped area}
\usage{
makeHDMValsfromRasters(myHDMSpp_NO = HDMSpp_NO, myCropRasters = cropRasters)
}
\arguments{
\item{myHDMSpp_NO}{vector of TAXON_IDs for species to be included in output}

\item{myCropRasters}{list of rasters and indices and cell values created by function cropNAborder()}
}
\description{
Function makes a matrix of HDM values(1,NA) constrained to those cells that
are indexed in the cropped area
}
