% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_abundance_functions.R
\name{makeSppYearSum2}
\alias{makeSppYearSum2}
\title{Species' relative abundance calculation and summary}
\usage{
makeSppYearSum2(
  FHAnalysis,
  myHDMSpp_NO = HDMSpp_NO,
  writeSpRasters = writeSpRasters,
  myLU_List = LU_List,
  ResultsDir = ResultsDir,
  HDMVals = HDMVals,
  TaxonList = TaxonList,
  writeYears = NULL,
  writeSp = writeSp
)
}
\arguments{
\item{FHAnalysis}{list containing all the fire history spatial attributes
created by function fhProcess}

\item{myHDMSpp_NO}{vector of TAXON_IDs for species to be included in output}

\item{writeSpRasters}{logical: whether to also write species abundance
rasters to disk}

\item{myLU_List}{list of species abundance lookup arrays created by function
make_Spp_LU_list()}

\item{ResultsDir}{path of directory where results will be written usually
generated  by FAME script}

\item{HDMVals}{matrix of cell values for Habitat Distribution Model rasters
for (at least) all TAXON_ID in myHDMSpp_NO generally provided in settings
file and read loaded by FAME script}

\item{TaxonList}{data.frame of species attributes ( read from default or user
provided .csv)}

\item{writeYears}{vector for SEASONS for which rasters are to be written
otherwise if writeSpRasters == TRUE, if writeYears == NULL then all SEASONS
are written out}

\item{writeSp}{vector of TAXON_IDs provided if only subset of species rasters
are required as output.}
}
\value{
data frame wide format summary of relative abundance of species by SEASONS
}
\description{
Species' relative abundance calculation and summary
}
\details{
Calculates the relative abundance of species for each raster cell in
analysis and summaries these as summed abundance each season. Optionally it
also write relative abundance rasters for species to disk
}
