% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDraftJFMP.R
\name{processDraftJFMP}
\alias{processDraftJFMP}
\title{Process draft JFMP inputs}
\usage{
processDraftJFMP(myDraftJFMPFile = rv$draftJFMPFile, myAutoJFMP = rv$autoJFMP)
}
\arguments{
\item{DraftJFMPFile}{path to csv file containing two columns:
PU  and Draft_JFMP_State which is the burn state
("Burn" or "NoBurn") for the planning unit(PU) in the draft JFMP
The PU values must be identical to those in myAutoJFMP}

\item{AutoJFMP}{autoJFMP for the corresponding fire History and
planning units}
}
\value{
data.frame JFMP table with one row for each planning unit,
and columns for:
– Planning unit ID (x1 column)
– FMZ category (x1)
– District (x1)
– Planning unit size in hectares (x1)
– Burn/non-burn state from draft JFMP (x 1)
– Score for each metric in burn/non-burn states (2 x 4)
– Difference in each metric between burn/non-burn states (x4)
– Ranking on difference in each metric between states within District (x4)
– Ranking as above but within District X FMZ combination (x 4)
This is the same as autoJFMP table  except different areas are considered for burning
}
\description{
Process draft JFMP inputs
}
\details{
Takes a table of burn units with identical PU ids to the autoJFMP can makes table of burn status for the draft JFMP and relevant ecological resilience and life and property risk scores per burn unit.
}
