% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterFHprocess.R
\name{raster_fh_process}
\alias{raster_fh_process}
\title{Raster Fire History Process}
\usage{
raster_fh_process(
  fireHistory = mySF,
  template_raster = template_r,
  OtherAndUnknown = 2
)
}
\arguments{
\item{template_raster}{raster providing the template extent, resolution etc
of output must be in same CRS and overlap the whole area of interest of the
input fire history polygon dataset}

\item{OtherAndUnknown}{integer Value to use for cases where fire type is:
"OTHER" or "UNKNOWN" = NA, "BURN" = 1, "BUSHFIRE" = 2. NA = Fire excluded
from analysis. usually set in settings file}
}
\description{
Raster Fire History Process
}
