% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFileModule.r
\name{selectFileUI}
\alias{selectFileUI}
\title{Module UI function for file selection}
\usage{
selectFileUI(id, label = "Select file", title = "", multiple)
}
\arguments{
\item{id}{string namespace id for module (same as in server)}

\item{label}{text to appear on select file button eg "Select file"}

\item{title}{text title for button (not used) default ""}

\item{multiple}{TRUE for allowing mutiple selection, default FALSE to select single file}
}
\value{
tibble of file attributes name, size, type, and path ( relative to root_dirs argument in fileSelectServer)
}
\description{
Module UI function for file selection
}
\details{
Select File Module UI used with selectFileServer
modified from https://stackoverflow.com/questions/38747129/how-to-use-shinyfiles-package-within-shiny-modules-namespace-issue
}
\examples{
simple file selection shiny app
ui <-   fluidPage(
selectFileUI(
  id = "fileSelect1",label ="New Label"
),
verbatimTextOutput("datapath")
)

server <- function(input, output, session) {
  rv<-reactiveValues()
  x <- selectFileServer(id = "fileSelect1")
  observe(rv$x<-x$datapathtab()$datapath)
  observeEvent(rv$x,output$datapath<-renderText(rv$x))

}
shinyApp(ui,server)


}
