% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadShapeFileGdbGpkg.r
\name{uploadShapeFileGdbGpkg}
\alias{uploadShapeFileGdbGpkg}
\title{Uploader for shapefiles zipped geodatabase or geopackage files to shiny app}
\usage{
uploadShapeFileGdbGpkg(
  myInput = "data frame of input files from shiny inputFiles",
  saveToPath = "path to save files to on server"
)
}
\arguments{
\item{myInput}{data frame of uploaded files selected using shiny inputFiles()
these files may be either:
a single zip file containing shapefile components (.shp
.shx,.prj,.dbf) and or one or more esri geodatabases (.gdb)
and or geopackage files
or:
a single geopackage file (.gpkg)
four files being  all four base components of a shapefile (.shp
.shx,.prj,.dbfp)}

\item{saveToPath}{path to save files to on server}
}
\value{
text message indicating success or otherwise of upload for rendering in shiny ui
}
\description{
Uploader for shapefiles zipped geodatabase or geopackage files to shiny app
}
