% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButtonDisableHelpers.r
\name{withBusyIndicatorUI}
\alias{withBusyIndicatorUI}
\title{Title withBusyIndicatorUI
ui part of withBusyIndicator module}
\usage{
withBusyIndicatorUI(button)
}
\arguments{
\item{button}{a shinyFilesButton object with an id}
}
\description{
Title withBusyIndicatorUI
ui part of withBusyIndicator module
}
