% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_raster_from_bin.R
\name{write_raster_from_bin}
\alias{write_raster_from_bin}
\title{Write a packed binary representation back to a raster}
\usage{
write_raster_from_bin(
  r_template,
  bin_file,
  index_file,
  out_raster_file,
  max_ncol,
  datatype = "INT2U",
  NAflag = 0,
  verbose = TRUE
)
}
\arguments{
\item{r_template}{SpatRaster template}

\item{bin_file}{Binary file from process_blocks_to_bin()}

\item{index_file}{Index CSV}

\item{out_raster_file}{Output GeoTIFF}

\item{max_ncol}{Number of layers in output}

\item{datatype}{Output GDAL datatype}

\item{NAflag}{NoData value}

\item{verbose}{Logical}
}
\description{
Write a packed binary representation back to a raster
}
