% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Spp_EFG_LMU.R
\name{calc_Spp_EFG_LMU}
\alias{calc_Spp_EFG_LMU}
\title{Calculate the species in each EFG in given area for GSO calculations.}
\usage{
calc_Spp_EFG_LMU(
  REG_NO,
  RasterRes = 225,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters",
  splist = "./ReferenceTables/DraftTaxonListStatewidev2.csv",
  myHDMVals = HDMVals225,
  myResultsDir = ResultsDir,
  TFI_LUT = TFI_LUT
)
}
\arguments{
\item{REG_NO}{integer DELWP fire region number 1:6 ,99 for Statewide analysis, or 7 for ad hoc boundary polygon default =7 (see look up table REG_LUT for values)}

\item{RasterRes}{integer 225 - raster resolution is always 225 for this function for speed}

\item{PUBLIC_LAND_ONLY}{logical whether to restrict analysis to public land only or the whole polygon}

\item{myPoly}{default clipPoly sf polygon data frame of LF_REGIONs (default) or ad hoc polygon - used in conjunction with REG_NO}

\item{generalRasterDir}{relative path to directory containing rasters of FIRE_REG, and PUBLIC LAND (PLM_GEN)}

\item{splist}{path to default species attribute table default is "./ReferenceTables/DraftTaxonListStatewidev2.csv"}

\item{myHDMVals}{matrix of cell values for Habitat Distribution Model rasters at 225m pixel size}

\item{myResultsDir}{path of directory where output will be saved}

\item{TFI_LUT}{data.frame lookup table for EFG loaded in setup}
}
\description{
works by using the indices of the standard dimensions raster that are in the
supplied shapefile region boundary (via function cropNAborder )
}
