% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBTFI_function.R
\name{calcBBTFI_2}
\alias{calcBBTFI_2}
\title{Calculate area BBTFI and BBTFI rasters}
\usage{
calcBBTFI_2(
  FHAnalysis,
  U_AllCombs_TFI = myAllCombs$U_AllCombs_TFI,
  Index_AllCombs = myAllCombs$Index_AllCombs,
  TFI_LUT,
  makeBBTFIrasters = makeBBTFIrasters
)
}
\arguments{
\item{FHAnalysis}{list containing all the fire history spatial attributes
created by function fhProcess}

\item{U_AllCombs_TFI}{data.table giving all combinations of cell values from
the input rasters for the FAME analysis}

\item{Index_AllCombs}{integer index mapping U_AllCombs_TFI to raster cells}

\item{TFI_LUT}{data.frame lookup table from EFG for
"MIN_LO_TFI","MIN_HI_TFI","MAX_TFI","EFG_NAME", read from settings}

\item{makeBBTFIrasters}{logical whether or not to export rasters for BBTFI to
disk}
}
\value{
list containing:
\itemize{
\item the date sequence matrix for each cell of the raster
\item the EFG TFI Lookup for each cell of the raster
\item the raster resolution used.
\item Optionally outputs rasters of BBTFI to disk if makeBBTFIrasters==TRUE.
}
}
\description{
Calculate area BBTFI and BBTFI rasters
}
\details{
Calculate summary  area burned below TFI BBTFI for each SEASON in analysis
(accommodating  Hi and Lo fire intensity of first burn to determine TFI)
and cumulative area BBTFI.  Also optionally outputs rasters mapping areas BBTFI
}
