% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDeltaAbund.R
\name{calcDeltaAbund}
\alias{calcDeltaAbund}
\title{Summary of changes in relative abundance}
\usage{
calcDeltaAbund(
  SpYearSumm = SpYearSummWide,
  myFHAnalysis,
  myBaseline,
  myResultsDir
)
}
\arguments{
\item{SpYearSumm}{data.frame output by function calc_SpeciesRA()}

\item{myFHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess}

\item{myBaseline}{integer single SEASON or sequence of SEASons used to create the baseline relative species abundance for comparison of change}

\item{myResultsDir}{path of directory where results will be written usually generated by FAME script
#'}
}
\value{
data frame wide format summary chance in relative abundance of species SEASON
}
\description{
Summary of changes in relative abundance
}
\details{
Calculates the change in relative abundance compared to a baseline SEASON or mean of SEASONS
}
