% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zeroToOne}
\alias{zeroToOne}
\title{linear rescale 0-1}
\usage{
zeroToOne(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{allows passing of na.rm}
}
\value{
numeric scale 0:1 of length(x)
}
\description{
linear rescale 0-1
}
\details{
Rescales numeric vector from 0:1 by dividing by difference between minimum and maximum values ignoring NA values. If x does not have two different numeric values then returns all zeros (ie all rank equal)
}
